# contracts/agent-internal-openapi.yaml
openapi: 3.1.0
info:
  title: Caijiatai Python Agent Internal API
  version: 0.5.0
  description: >-
    Internal-only API used by the Java procurement control plane.
    Production transport is Kafka RPC: commands arrive on topic caijiatai.commands as
    HMAC-SHA256 signed envelopes (MessageCodec.signEnvelope; signature over the
    canonical JSON excluding the signature field itself) and results return on
    caijiatai.results; both sides verify signature plus payload_sha256. The HTTP
    surface documented here is the legacy loopback mode: every /internal/v1/* path
    (and, in internal mode, every non-health path) is gated by the
    X-Agent-Internal-Token header with a constant-time compare; POST is only allowed
    for /internal/v1/commands and /internal/v1/config.
servers:
  - url: http://agent:8742
security:
  - InternalToken: []
paths:
...
  /internal/v1/commands:
    post:
      operationId: executeAgentCommand
      parameters:
        - $ref: '#/components/parameters/OperationId'
        - $ref: '#/components/parameters/PayloadSha256'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCommand'
      responses:
        '200':
          description: Existing or newly completed idempotent result.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AgentCommandResult'}
        '202':
          description: Command accepted and still running.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AgentCommandResult'}
        '409':
          description: Same operation id was previously used with a different payload.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  /internal/v1/config:
    get:
      operationId: getModelConfig
      responses:
        '200':
          description: >-
            Effective procurement model config. The raw API key is never returned;
            since v0.5.1 the response carries only the api_key_configured boolean
            (api_key_preview removed).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ModelConfig'}
    post:
      operationId: updateModelConfig
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ModelConfigUpdate'}
      responses:
        '200':
          description: Persisted config echoed back in read form.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ModelConfig'}
        '422':
          description: Unknown fields, invalid provider/planner_mode, blank model, or bad price fields.
...
  /api/runs/{runId}:
    get:
      operationId: getRun
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Run, content: {application/json: {schema: {type: object}}}}
  /api/runs/{runId}/report:
    get:
      operationId: getRunReport
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Runtime report, content: {application/json: {schema: {type: object}}}}
  /api/runs/{runId}/messages:
    get:
      operationId: getRunMessages
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Run messages, content: {application/json: {schema: {type: array, items: {type: object}}}}}
  /api/runs/{runId}/events:
    get:
      operationId: getRunEvents
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Run events, content: {application/json: {schema: {type: array, items: {type: object}}}}}
  /api/runs/{runId}/approvals:
    get:
      operationId: getRunApprovals
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Run approvals, content: {application/json: {schema: {type: array, items: {type: object}}}}}
  /api/runs/{runId}/tool-invocations:
    get:
      operationId: getRunToolInvocations
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Tool invocations, content: {application/json: {schema: {type: array, items: {type: object}}}}}
  /api/runs/{runId}/checkpoint:
    get:
      operationId: getRunCheckpoint
      parameters: [{$ref: '#/components/parameters/RunId'}]
      responses:
        '200': {description: Checkpoint, content: {application/json: {schema: {type: object}}}}
...
  /api/sessions/{sessionId}/transcript:
    get:
      operationId: getSessionTranscript
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: {type: string, pattern: '^[0-9a-f]{32}$'}
      responses:
        '200': {description: Transcript, content: {application/json: {schema: {type: object}}}}
  /api/tool-invocations/{invocationId}:
    get:
      operationId: getToolInvocation
      parameters:
        - name: invocationId
          in: path
          required: true
          schema: {type: string, pattern: '^[0-9a-f]{32}$'}
      responses:
        '200': {description: Tool invocation, content: {application/json: {schema: {type: object}}}}
components:
  securitySchemes:
    InternalToken:
      type: apiKey
      in: header
      name: X-Agent-Internal-Token
      description: >-
        Legacy HTTP loopback mode only. In production Kafka RPC mode there is no
        HTTP hop between Java and Python; authenticity comes from the HMAC-SHA256
        envelope signature (shared internal HMAC key) plus payload_sha256 checks.
  parameters:
    OperationId:
      name: X-Operation-Id
      in: header
      required: true
      schema: {type: string, format: uuid}
    PayloadSha256:
      name: X-Payload-SHA256
      in: header
      required: true
      schema: {type: string, pattern: '^[0-9a-f]{64}$'}
    RunId:
      name: runId
      in: path
      required: true
      schema: {type: string, pattern: '^[0-9a-f]{32}$'}
  schemas:
    AgentCommand:
      type: object
      additionalProperties: false
      required: [operation_id, operation_type, aggregate_id, generation, expected_task_version, payload_sha256, payload]
      properties:
        operation_id: {type: string, format: uuid}
        operation_type: {type: string, enum: [start_conversation, import_quote, resume_run, human_interaction_answer, analyze, approve_decision, create_structured, reopen_task, parse_invoice, explain_invoice_diff, draft_contract]}
        aggregate_id: {type: string, pattern: '^[0-9a-f]{32}$'}
        generation: {type: integer, minimum: 1}
        expected_task_version: {type: integer, minimum: 0}
        payload_sha256: {type: string, pattern: '^[0-9a-f]{64}$'}
        payload: {type: object}
    AgentCommandResult:
      type: object
      required: [operation_id, status]
      properties:
        operation_id: {type: string, format: uuid}
        status: {type: string, enum: [accepted, completed, failed]}
        result: {type: [object, 'null']}
        error: {type: [string, 'null']}
    ErrorResponse:
      type: object
      required: [code, message, request_id, status]
      properties:
        code: {type: string}
        message: {type: string}
        request_id: {type: string, format: uuid}
        status: {type: integer}
    ModelConfig:
      type: object
      description: >-
        Read form of the procurement model config. api_key/api_key_from_env are
        internal persistence fields and are stripped from every response.
      required: [provider, planner_mode, model, api_mode, reasoning_effort, api_key_configured]
      properties:
        provider: {type: string, enum: [procurement_fake, openai]}
        planner_mode: {type: string, enum: [model, hybrid]}
        model: {type: string}
        base_url: {type: [string, 'null']}
        api_mode: {type: string, enum: [auto, responses, chat]}
        reasoning_effort: {type: string}
        api_key_configured: {type: boolean}
        input_price_per_million_usd: {type: [number, 'null']}
        output_price_per_million_usd: {type: [number, 'null']}
        cached_input_price_per_million_usd: {type: [number, 'null']}
        max_cost_usd: {type: [number, 'null']}
    ModelConfigUpdate:
      type: object
      description: >-
        Write payload for POST /internal/v1/config. A supplied api_key is persisted
        only when it differs from OPENAI_API_KEY (env-matching keys are stored as an
        api_key_from_env reference, never in plaintext).
      additionalProperties: false
      required: [provider, model]
      properties:
        provider: {type: string, enum: [procurement_fake, openai]}
        planner_mode: {type: string, enum: [model, hybrid]}
        model: {type: string, minLength: 1}
        base_url: {type: [string, 'null']}
        api_mode: {type: string, enum: [auto, responses, chat]}
        reasoning_effort: {type: string}
        api_key: {type: [string, 'null']}
        input_price_per_million_usd: {type: [number, 'null']}
        output_price_per_million_usd: {type: [number, 'null']}
        cached_input_price_per_million_usd: {type: [number, 'null']}
        max_cost_usd: {type: [number, 'null']}